added footracegame base and saved old one
